Refactor/user route use magic (#7)
* feature(auth-controller): detect reused refresh token

* chore(auth-route): replace comment for a todo feature

* fix(profile-service): check if there is avatar before deleting

* refactor(block-user-service): remove redundant block user check

* refactor(user): use magic ID instead of user ID

+++ src/v1/user/profile/profile-service.js
@@ -202,6 +202,10 @@ const createDeleteProfileAvatarByUserId =
 
     const profile = await profileRepository.findProfileByUserPk(userPk);
 
+    if (profile.avatar === null) {
+      throw new APIError("No avatar found", httpStatus.NOT_FOUND);
+    }
+
     await storage.destroyFile(profile.avatar.id, "Image");
 
     return profileRepository.deleteAvatarByUserPk(userPk);
@@ -218,7 +222,9 @@ const createDeleteBackgroundAvatarByUserId =
 
     const profile = await profileRepository.findProfileByUserPk(userPk);
 
-    await storage.destroyFile(profile.backgroundAvatar.id, "Image");
+    if (profile.backgroundAvatar === null) {
+      throw new APIError("No background avatar found", httpStatus.NOT_FOUND);
+    }
 
     await storage.destroyFile(profile.backgroundAvatar.id, "Image");
